Added stdint.h and inttypes.h libraries

+++ chapter6/ProgrammingProjects/Q2-GreatestCommonDivisor/GCFHead.h
@@ -1,10 +1,13 @@
 #ifndef HEADER 
 #define HEADER 
 
+#include <stdint.h>
+#include <inttypes.h>
+
 void print_tokens(char* firstToken, char* secondToken);
 
 int gcf_table(int argc, char* argv[]);
 
-long long calculate_gcf(long long n0, long long n1);
+uint64_t calculate_gcf(uint64_t n0, uint64_t n1);
 
 #endif
